Move gitolite stuff into own file
